Refactor context usage and add new endpoints

+++ src/AcoTerra.API/Data/Entities/Products/Product.cs
@@ -9,7 +9,7 @@ internal sealed class Product : AuditableEntity
     public int Id { get; set; }
     public required string Name { get; set; }
     public required PackagingType PackagingType { get; set; }
-    public required double Weight { get; set; }
+    public required decimal Weight { get; set; }
     public required decimal Price { get; set; }
 }
